wait for cluster readiness before each operation Fixes #2691
