--- extensions-jvm/couchbase/integration-test/src/test/java/org/apache/camel/quarkus/component/couchbase/it/CouchbaseDeleteTest.java
@@ -1,53 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.camel.quarkus.component.couchbase.it;
-
-import io.quarkus.test.common.QuarkusTestResource;
-import io.quarkus.test.common.http.TestHTTPEndpoint;
-import io.quarkus.test.junit.QuarkusTest;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.condition.DisabledIfEnvironmentVariable;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.CoreMatchers.equalTo;
-
-@QuarkusTest
-@TestHTTPEndpoint(CouchbaseResource.class)
-@QuarkusTestResource(CouchbaseTestResource.class)
-@DisabledIfEnvironmentVariable(named = "CI", matches = "true")
-class CouchbaseDeleteTest {
-
-    @Test
-    void testDelete() {
-        // getting the document
-        given()
-                .when()
-                .get("/DocumentID_1")
-                .then()
-                .statusCode(200)
-                .body(equalTo("hello1"));
-
-        // deleting the document
-        given()
-                .when()
-                .delete("DocumentID_1")
-                .then()
-                .statusCode(200);
-
-    }
-
-}
